remove snapshottest from the example tests
so that the tests pass in 3.12 and 3.13 again
