+++ examples/starwars/tests/test_query.py
@@ -8,19 +8,19 @@ setup()
 client = Client(schema)
 
 
-def test_hero_name_query(snapshot):
-    query = """
+def test_hero_name_query():
+    result = client.execute("""
         query HeroNameQuery {
           hero {
             name
           }
         }
-    """
-    snapshot.assert_match(client.execute(query))
+    """)
+    assert result == {"data": {"hero": {"name": "R2-D2"}}}
 
 
-def test_hero_name_and_friends_query(snapshot):
-    query = """
+def test_hero_name_and_friends_query():
+    result = client.execute("""
         query HeroNameAndFriendsQuery {
           hero {
             id
@@ -30,12 +30,24 @@ def test_hero_name_and_friends_query(snapshot):
             }
           }
         }
-    """
-    snapshot.assert_match(client.execute(query))
+    """)
+    assert result == {
+        "data": {
+            "hero": {
+                "id": "2001",
+                "name": "R2-D2",
+                "friends": [
+                    {"name": "Luke Skywalker"},
+                    {"name": "Han Solo"},
+                    {"name": "Leia Organa"},
+                ],
+            }
+        }
+    }
 
 
-def test_nested_query(snapshot):
-    query = """
+def test_nested_query():
+    result = client.execute("""
         query NestedQuery {
           hero {
             name
@@ -48,70 +60,113 @@ def test_nested_query(snapshot):
             }
           }
         }
-    """
-    snapshot.assert_match(client.execute(query))
+    """)
+    assert result == {
+        "data": {
+            "hero": {
+                "name": "R2-D2",
+                "friends": [
+                    {
+                        "name": "Luke Skywalker",
+                        "appearsIn": ["NEWHOPE", "EMPIRE", "JEDI"],
+                        "friends": [
+                            {"name": "Han Solo"},
+                            {"name": "Leia Organa"},
+                            {"name": "C-3PO"},
+                            {"name": "R2-D2"},
+                        ],
+                    },
+                    {
+                        "name": "Han Solo",
+                        "appearsIn": ["NEWHOPE", "EMPIRE", "JEDI"],
+                        "friends": [
+                            {"name": "Luke Skywalker"},
+                            {"name": "Leia Organa"},
+                            {"name": "R2-D2"},
+                        ],
+                    },
+                    {
+                        "name": "Leia Organa",
+                        "appearsIn": ["NEWHOPE", "EMPIRE", "JEDI"],
+                        "friends": [
+                            {"name": "Luke Skywalker"},
+                            {"name": "Han Solo"},
+                            {"name": "C-3PO"},
+                            {"name": "R2-D2"},
+                        ],
+                    },
+                ],
+            }
+        }
+    }
 
 
-def test_fetch_luke_query(snapshot):
-    query = """
+def test_fetch_luke_query():
+    result = client.execute("""
         query FetchLukeQuery {
           human(id: "1000") {
             name
           }
         }
-    """
-    snapshot.assert_match(client.execute(query))
+    """)
+    assert result == {"data": {"human": {"name": "Luke Skywalker"}}}
 
 
-def test_fetch_some_id_query(snapshot):
-    query = """
+def test_fetch_some_id_query():
+    result = client.execute(
+        """
         query FetchSomeIDQuery($someId: String!) {
           human(id: $someId) {
             name
           }
         }
-    """
-    params = {"someId": "1000"}
-    snapshot.assert_match(client.execute(query, variables=params))
+    """,
+        variables={"someId": "1000"},
+    )
+    assert result == {"data": {"human": {"name": "Luke Skywalker"}}}
 
 
-def test_fetch_some_id_query2(snapshot):
-    query = """
+def test_fetch_some_id_query2():
+    result = client.execute(
+        """
         query FetchSomeIDQuery($someId: String!) {
           human(id: $someId) {
             name
           }
         }
-    """
-    params = {"someId": "1002"}
-    snapshot.assert_match(client.execute(query, variables=params))
+    """,
+        variables={"someId": "1002"},
+    )
+    assert result == {"data": {"human": {"name": "Han Solo"}}}
 
 
-def test_invalid_id_query(snapshot):
-    query = """
+def test_invalid_id_query():
+    result = client.execute(
+        """
         query humanQuery($id: String!) {
           human(id: $id) {
             name
           }
         }
-    """
-    params = {"id": "not a valid id"}
-    snapshot.assert_match(client.execute(query, variables=params))
+    """,
+        variables={"id": "not a valid id"},
+    )
+    assert result == {"data": {"human": None}}
 
 
-def test_fetch_luke_aliased(snapshot):
-    query = """
+def test_fetch_luke_aliased():
+    result = client.execute("""
         query FetchLukeAliased {
           luke: human(id: "1000") {
             name
           }
         }
-    """
-    snapshot.assert_match(client.execute(query))
+    """)
+    assert result == {"data": {"luke": {"name": "Luke Skywalker"}}}
 
 
-def test_fetch_luke_and_leia_aliased(snapshot):
-    query = """
+def test_fetch_luke_and_leia_aliased():
+    result = client.execute("""
         query FetchLukeAndLeiaAliased {
           luke: human(id: "1000") {
             name
@@ -120,12 +175,14 @@ def test_fetch_luke_and_leia_aliased(snapshot):
             name
           }
         }
-    """
-    snapshot.assert_match(client.execute(query))
+    """)
+    assert result == {
+        "data": {"luke": {"name": "Luke Skywalker"}, "leia": {"name": "Leia Organa"}}
+    }
 
 
-def test_duplicate_fields(snapshot):
-    query = """
+def test_duplicate_fields():
+    result = client.execute("""
         query DuplicateFields {
           luke: human(id: "1000") {
             name
@@ -136,12 +193,17 @@ def test_duplicate_fields(snapshot):
             homePlanet
           }
         }
-    """
-    snapshot.assert_match(client.execute(query))
+    """)
+    assert result == {
+        "data": {
+            "luke": {"name": "Luke Skywalker", "homePlanet": "Tatooine"},
+            "leia": {"name": "Leia Organa", "homePlanet": "Alderaan"},
+        }
+    }
 
 
-def test_use_fragment(snapshot):
-    query = """
+def test_use_fragment():
+    result = client.execute("""
         query UseFragment {
           luke: human(id: "1000") {
             ...HumanFragment
@@ -154,29 +216,36 @@ def test_use_fragment(snapshot):
           name
           homePlanet
         }
-    """
-    snapshot.assert_match(client.execute(query))
+    """)
+    assert result == {
+        "data": {
+            "luke": {"name": "Luke Skywalker", "homePlanet": "Tatooine"},
+            "leia": {"name": "Leia Organa", "homePlanet": "Alderaan"},
+        }
+    }
 
 
-def test_check_type_of_r2(snapshot):
-    query = """
+def test_check_type_of_r2():
+    result = client.execute("""
         query CheckTypeOfR2 {
           hero {
             __typename
             name
           }
         }
-    """
-    snapshot.assert_match(client.execute(query))
+    """)
+    assert result == {"data": {"hero": {"__typename": "Droid", "name": "R2-D2"}}}
 
 
-def test_check_type_of_luke(snapshot):
-    query = """
+def test_check_type_of_luke():
+    result = client.execute("""
         query CheckTypeOfLuke {
           hero(episode: EMPIRE) {
             __typename
             name
           }
         }
-    """
-    snapshot.assert_match(client.execute(query))
+    """)
+    assert result == {
+        "data": {"hero": {"__typename": "Human", "name": "Luke Skywalker"}}
+    }
