+++ examples/starwars_relay/tests/test_mutation.py
@@ -8,26 +8,45 @@ setup()
 client = Client(schema)
 
 
-def test_mutations(snapshot):
-    query = """
-    mutation MyMutation {
-      introduceShip(input:{clientMutationId:"abc", shipName: "Peter", factionId: "1"}) {
-        ship {
-          id
-          name
-        }
-        faction {
-          name
-          ships {
-            edges {
-              node {
-                id
-                name
+def test_mutations():
+    result = client.execute("""
+        mutation MyMutation {
+          introduceShip(input:{clientMutationId:"abc", shipName: "Peter", factionId: "1"}) {
+            ship {
+              id
+              name
+            }
+            faction {
+              name
+              ships {
+                edges {
+                  node {
+                    id
+                    name
+                  }
+                }
               }
             }
           }
         }
-      }
+    """)
+    assert result == {
+        "data": {
+            "introduceShip": {
+                "ship": {"id": "U2hpcDo5", "name": "Peter"},
+                "faction": {
+                    "name": "Alliance to Restore the Republic",
+                    "ships": {
+                        "edges": [
+                            {"node": {"id": "U2hpcDox", "name": "X-Wing"}},
+                            {"node": {"id": "U2hpcDoy", "name": "Y-Wing"}},
+                            {"node": {"id": "U2hpcDoz", "name": "A-Wing"}},
+                            {"node": {"id": "U2hpcDo0", "name": "Millennium Falcon"}},
+                            {"node": {"id": "U2hpcDo1", "name": "Home One"}},
+                            {"node": {"id": "U2hpcDo5", "name": "Peter"}},
+                        ]
+                    },
+                },
+            }
+        }
     }
-    """
-    snapshot.assert_match(client.execute(query))
